feat(core): :art: improve last refactoring
