--- src/main/java/com/sevenmap/data/objsept/Convertor.java
@@ -1,5 +1,0 @@
-package com.sevenmap.data.objsept;
-
-public class Convertor {
-
-}
