+++ src/main/java/com/sevenmap/data/objsept/Portal.java
@@ -1,4 +1,4 @@
-package com.sevenmap.data.objset;
+package com.sevenmap.data.objsept;
 
 import java.util.ArrayList;
 import java.util.List;
